webdemo/requirements.txt
Flask
gunicorn
joblib
numpy
pandas
gdown
matplotlib
scikit-learn
tqdm
torch
torchvision
pillow
